+++ lib/ZuluIDE_platform_RP2040/display_ssd1306.h
@@ -71,25 +71,6 @@ namespace zuluide {
 0x0,0xf,0xe0,0x0,
 0x0,0x0,0x0,0x0
 };
-// {
-// 0x0,0x0,0x0,0x0,
-// 0x0,0xf,0xf0,0x0,
-// 0x0,0xff,0xff,0x0,
-// 0x3,0xff,0xff,0xc0,
-// 0xf,0xff,0xff,0xf0,
-// 0x1f,0xfc,0x3f,0xf8,
-// 0x3f,0xf9,0x9f,0xfc,
-// 0x3f,0xfb,0xdf,0xfc,
-// 0x3f,0xf9,0x9f,0xfc,
-// 0x3f,0xfc,0x3f,0xfc,
-// 0x1f,0xff,0xff,0xf8,
-// 0xf,0xff,0xff,0xf0,
-// 0x3,0xff,0xff,0xc0,
-// 0x0,0xff,0xff,0x0,
-// 0x0,0xf,0xf0,0x0,
-// 0x0,0x0,0x0,0x0
-// };
-
 
     uint8_t cdrom_empty[16*4] =
 {
